New workflows, some more documentation, and a relaunch helper!
The new relaunch helper is absolutely awesome, this is the first time I've ever done anything like this.
I hope its not bad code, I tried to make sure this was safe from any possible segfaults since I wrote it in Objective-C. It should work on 10.13 but we'll have to see that at a later time.